Adds Screen2
