+++ code/Controller/data/buttons.ts
@@ -1,0 +1,22 @@
+enum Buttons {
+  Safe = 'Safe',
+  Drill = 'Drill',
+  SIP = 'Shelter in Place',
+  Exit = 'Exit',
+}
+
+enum ButtonColors {
+  Green = 'Green',
+  Blue = 'Blue',
+  Yellow = 'Yellow',
+  Red = 'Red',
+}
+
+const ColorKeys = {
+  Green: '#00FF00',
+  Blue: '#5555FF',
+  Yellow: '#BBBB00',
+  Red: '#FF0000',
+};
+
+export {Buttons, ButtonColors, ColorKeys};
