+++ code/Controller/package.json
@@ -10,8 +10,13 @@
     "test": "jest"
   },
   "dependencies": {
+    "@react-navigation/native": "^6.1.9",
+    "@react-navigation/native-stack": "^6.9.17",
     "react": "18.2.0",
-    "react-native": "0.72.7"
+    "react-native": "0.72.7",
+    "react-native-gesture-handler": "^2.14.0",
+    "react-native-safe-area-context": "^4.7.4",
+    "react-native-screens": "^3.27.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
